app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-api-python-client==2.141.0
google-auth==2.34.0
python-multipart==0.0.9
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0